refactor: 드래그 및 터치 이벤트 핸들러 모듈로 분리
