+++ src/components/Carousel.js
@@ -1,6 +1,5 @@
-import { getClientCoordinate } from '../utils/coordinateUtils.js';
+import { addDragAndTouchEventHandlers } from '../utils/addDragAndTouchEventHandlers.js';
 
-const SWIPE_DISTANCE_THRESHOLD_IN_PIXEL = 50; // 스와이프 임계값
 const DEFAULT_SLIDE_SPEED = 55_000; // 기본 슬라이드 속도
 
 export function Carousel({ slides, carouselSlide, delay, transitionSpeed }) {
@@ -12,11 +11,6 @@ export function Carousel({ slides, carouselSlide, delay, transitionSpeed }) {
 
   this.currentSlide = 0;
   this.isPlaying = true;
-  let startX = 0,
-    startY = 0,
-    endX = 0,
-    endY = 0,
-    isDragging = false;
 
   const indicatorButtons = [];
 
@@ -159,58 +153,8 @@ export function Carousel({ slides, carouselSlide, delay, transitionSpeed }) {
     this.autoFlipSlide(delay + transitionSpeed);
   };
 
-  // 마우스 드래그 및 터치 무브 이벤트 핸들러
-  const handleDragAndTouchMoveEvents = () => {
-    const handlePointerDown = (event) => {
-      const { clientX, clientY } = getClientCoordinate(event);
+  this.render(); // 초기 렌더링
+  addDragAndTouchEventHandlers(carouselSlide, this.navigateSlide); // 터치 이벤트, 드래그 이벤트 등록
 
-      startX = clientX;
-      startY = clientY;
-      isDragging = true;
-      event.preventDefault();
-    };
-
-    const handlePointerMove = (event) => {
-      if (!isDragging) return;
-      const { clientX, clientY } = getClientCoordinate(event);
-
-      endX = clientX;
-      endY = clientY;
-    };
-
-    const handlePointerUp = () => {
-      if (!isDragging) return;
-
-      const swipeDistanceX = endX - startX;
-      const swipeDistanceY = endY - startY;
-
-      const angle = Math.abs(swipeDistanceY / swipeDistanceX);
-
-      // 이동 거리가 50px이상 && 이동 각도가 45도 미만일 때만 드래그
-      if (
-        Math.abs(swipeDistanceX) > SWIPE_DISTANCE_THRESHOLD_IN_PIXEL &&
-        angle < 1
-      ) {
-        this.navigateSlide(swipeDistanceX > 0 ? 'prev' : 'next');
-      }
-
-      isDragging = false;
-    };
-
-    carouselSlide.addEventListener('mousedown', handlePointerDown);
-    carouselSlide.addEventListener('mousemove', handlePointerMove);
-    carouselSlide.addEventListener('mouseup', handlePointerUp);
-    carouselSlide.addEventListener('mouseleave', handlePointerUp); // 캐러셀 바깥으로 나갈 때 드래그 종료하도록
-
-    carouselSlide.addEventListener('touchstart', handlePointerDown);
-    carouselSlide.addEventListener('touchmove', handlePointerMove);
-    carouselSlide.addEventListener('touchend', handlePointerUp);
-  };
-
-  // 초기 렌더링 및 터치 이벤트, 드래그 이벤트 등록
-  this.render();
-  handleDragAndTouchMoveEvents();
-
-  // 페이지가 닫힐 때 타이머 제거
-  window.addEventListener('beforeunload', this.clearAutoFlipSlide);
+  window.addEventListener('beforeunload', this.clearAutoFlipSlide); // 페이지가 닫힐 때 타이머 제거
 }
